refactor: modularize download functionality and format evaluation results
- Extracted the evaluation formatting logic into a separate utility function for better code organization and reusability.
- Created a new utility function for downloading text files, simplifying the download process in the MockInterviewEvaluation component.
- Updated the component to utilize these new utilities, enhancing maintainability and clarity.
