+++ components/mock-interview/MockInterviewEvaluation.tsx
@@ -3,34 +3,11 @@ import EvaluationCard from '../EvaluationCard';
 import { EvaluationResult } from '@/types/Evaluation';
 import FuturisticButton from '../FuturisticButton';
 import { FaDownload } from 'react-icons/fa';
+import { formatEvaluationForDownload } from '@/utils/functions/formatEvaluationForDownload';
+import { downloadTextFile } from '@/utils/functions/downloadTextFile';
 
 interface MockInterviewEvaluationProps {
   evaluation: EvaluationResult;
-}
-
-function formatEvaluationForDownload(evaluation: EvaluationResult) {
-  let content = `Mock Interview Result\n\n`;
-  evaluation.results.forEach((result, idx) => {
-    content += `Question ${idx + 1}:\n`;
-    content += `${result.question}\n\n`;
-    content += `Key Points:\n`;
-    if (
-      result.key_points_of_main_argument &&
-      result.key_points_of_main_argument.length > 0
-    ) {
-      result.key_points_of_main_argument.forEach((kp) => {
-        content += `  - ${kp}\n`;
-      });
-    } else {
-      content += `  (None)\n`;
-    }
-    content += `\nPerfect Answer:\n`;
-    content += result.perfect_answer
-      ? `${result.perfect_answer}\n`
-      : '(None)\n';
-    content += `\n-----------------------------\n\n`;
-  });
-  return content;
 }
 
 const MockInterviewEvaluation: React.FC<MockInterviewEvaluationProps> = ({
@@ -39,17 +16,10 @@ const MockInterviewEvaluation: React.FC<MockInterviewEvaluationProps> = ({
   const evaluationRef = useRef(null);
   const handleDownload = () => {
     const content = formatEvaluationForDownload(evaluation);
-    const blob = new Blob([content], { type: 'text/plain' });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = `mock-interview-result-${
+    const filename = `mock-interview-result-${
       new Date().toISOString().split('T')[0]
     }.txt`;
-    document.body.appendChild(a);
-    a.click();
-    document.body.removeChild(a);
-    URL.revokeObjectURL(url);
+    downloadTextFile(content, filename);
   };
 
   return (
